+++ src/components/containers/Sidebar/index.tsx
@@ -4,13 +4,13 @@ import { useTranslation } from 'react-i18next';
 import { useSetRecoilState } from 'recoil';
 import { useAuth } from '@libs/auth';
 
-import TotalLectureSWListTab from '@components/containers/TotalLectureSWListTab';
-import AddOrUpdateLectureSWTab from '@components/containers/AddOrUpdateLectureSWTab';
-import SubscribedSWTab from '@components/containers/SubscribedSWTab';
-import LectureSWManagementTab from '@components/containers/LectureSWManagementTab';
-import ProjectListTab from '@components/containers/ProjectListTab';
-import AddOrUpdateProjectTab from '@components/containers/AddOrUpdateProjectTab';
-import LicenseListTab from '@components/containers/LicenseListTab';
+import TotalLectureSWListTab from '@components/tabs/TotalLectureSWListTab';
+import AddOrUpdateLectureSWTab from '@components/tabs/AddOrUpdateLectureSWTab';
+import SubscribedSWTab from '@components/tabs/SubscribedSWTab';
+import LectureSWManagementTab from '@components/tabs/LectureSWManagementTab';
+import ProjectListTab from '@components/tabs/ProjectListTab';
+import AddOrUpdateProjectTab from '@components/tabs/AddOrUpdateProjectTab';
+import LicenseListTab from '@components/tabs/LicenseListTab';
 import Dropdown from '@components/widgets/Dropdown';
 import Accordion from '@components/widgets/Accordion';
 import UserInfo from './UserInfo';
