+++ src/components/tabs/LectureSWManagementTab/LectureSWManagementTab.stories.tsx
