+++ src/components/tabs/LectureSWManagementTab/styled.ts
