+++ src/components/tabs/TotalLectureSWListTab/styled.ts
